+++ gazebo/gui/Actions.hh
@@ -38,7 +38,6 @@ namespace gazebo
 
     extern QAction *g_dataLoggerAct;
 
-    extern QAction *g_newModelAct;
     extern QAction *g_resetModelsAct;
     extern QAction *g_resetWorldAct;
     extern QAction *g_editBuildingAct;
@@ -88,6 +87,7 @@ namespace gazebo
     extern QAction *g_pasteAct;
 
     extern QWidgetAction *g_alignAct;
+    extern QAction *g_alignButtonAct;
     extern QAction *g_snapAct;
 
     /// \class DeleteAction Actions.hh gui/gui.hh