+++ gazebo/gui/GLWidget.cc
@@ -147,6 +147,9 @@ GLWidget::GLWidget(QWidget *_parent)
 
   connect(g_copyAct, SIGNAL(triggered()), this, SLOT(OnCopy()));
   connect(g_pasteAct, SIGNAL(triggered()), this, SLOT(OnPaste()));
+
+  connect(g_editModelAct, SIGNAL(toggled(bool)), this,
+      SLOT(OnModelEditor(bool)));
 }
 
 /////////////////////////////////////////////////
@@ -291,20 +294,26 @@ void GLWidget::keyPressEvent(QKeyEvent *_event)
     }
   }
 
-  this->mouseEvent.control =
+  this->keyEvent.control =
     this->keyModifiers & Qt::ControlModifier ? true : false;
-  this->mouseEvent.shift =
+  this->keyEvent.shift =
     this->keyModifiers & Qt::ShiftModifier ? true : false;
-  this->mouseEvent.alt =
+  this->keyEvent.alt =
     this->keyModifiers & Qt::AltModifier ? true : false;
+
+  this->mouseEvent.control = this->keyEvent.control;
+  this->mouseEvent.shift = this->keyEvent.shift;
+  this->mouseEvent.alt = this->keyEvent.alt;
 
   if (this->mouseEvent.control)
   {
-    if (_event->key() == Qt::Key_C && !this->selectedVisuals.empty())
+    if (_event->key() == Qt::Key_C && !this->selectedVisuals.empty()
+       && !this->modelEditorEnabled)
     {
       g_copyAct->trigger();
     }
-    else if (_event->key() == Qt::Key_V && !this->copyEntityName.empty())
+    else if (_event->key() == Qt::Key_V && !this->copyEntityName.empty()
+       && !this->modelEditorEnabled)
     {
       g_pasteAct->trigger();
     }
@@ -351,13 +360,16 @@ void GLWidget::keyReleaseEvent(QKeyEvent *_event)
       g_arrowAct->trigger();
   }
 
-  this->mouseEvent.control =
+  this->keyEvent.control =
     this->keyModifiers & Qt::ControlModifier ? true : false;
-  this->mouseEvent.shift =
+  this->keyEvent.shift =
     this->keyModifiers & Qt::ShiftModifier ? true : false;
-  this->mouseEvent.alt =
+  this->keyEvent.alt =
     this->keyModifiers & Qt::AltModifier ? true : false;
 
+  this->mouseEvent.control = this->keyEvent.control;
+  this->mouseEvent.shift = this->keyEvent.shift;
+  this->mouseEvent.alt = this->keyEvent.alt;
 
   ModelManipulator::Instance()->OnKeyReleaseEvent(this->keyEvent);
   this->keyText = "";
@@ -985,14 +997,17 @@ void GLWidget::OnManipMode(const std::string &_mode)
 /////////////////////////////////////////////////
 void GLWidget::OnCopy()
 {
-  if (!this->selectedVisuals.empty())
+  if (!this->selectedVisuals.empty() && !this->modelEditorEnabled)
+  {
     this->Copy(this->selectedVisuals.back()->GetName());
+  }
 }
 
 /////////////////////////////////////////////////
 void GLWidget::OnPaste()
 {
-  this->Paste(this->copyEntityName);
+  if (!this->modelEditorEnabled)
+    this->Paste(this->copyEntityName);
 }
 
 /////////////////////////////////////////////////
@@ -1138,4 +1153,19 @@ void GLWidget::OnAlignMode(const std::string &_axis, const std::string &_config,
 {
   ModelAlign::Instance()->AlignVisuals(this->selectedVisuals, _axis, _config,
       _target, !_preview);
+}
+
+/////////////////////////////////////////////////
+void GLWidget::OnModelEditor(bool _checked)
+{
+  this->modelEditorEnabled = _checked;
+  g_arrowAct->trigger();
+  event::Events::setSelectedEntity("", "normal");
+
+  // Manually deselect, in case the editor was opened with Ctrl
+  for (unsigned int i = 0; i < this->selectedVisuals.size(); ++i)
+  {
+    this->selectedVisuals[i]->SetHighlighted(false);
+  }
+  this->selectedVisuals.clear();
 }