+++ gazebo/gui/RenderWidget.cc
@@ -96,7 +96,7 @@ RenderWidget::RenderWidget(QWidget *_parent)
   alignMenu->addAction(g_alignAct);
   alignButton->setMenu(alignMenu);
   alignButton->setPopupMode(QToolButton::InstantPopup);
-  this->toolbar->addWidget(alignButton);
+  g_alignButtonAct = this->toolbar->addWidget(alignButton);
   connect(alignButton, SIGNAL(pressed()), g_alignAct, SLOT(trigger()));
 
   this->toolbar->addSeparator();
@@ -297,6 +297,7 @@ void RenderWidget::ShowEditor(bool _show)
     this->baseOverlayMsg = "Building is view-only";
     this->OnClearOverlayMsg();
     this->bottomFrame->hide();
+    this->ShowToolbar(false);
   }
   else
   {
@@ -304,6 +305,7 @@ void RenderWidget::ShowEditor(bool _show)
     this->baseOverlayMsg = "";
     this->OnClearOverlayMsg();
     this->bottomFrame->show();
+    this->ShowToolbar(true);
   }
 }