+++ gazebo/gui/building/BuildingModelManip.cc
@@ -18,6 +18,7 @@
 #include "gazebo/rendering/Visual.hh"
 #include "gazebo/common/Exception.hh"
 #include "gazebo/math/Quaternion.hh"
+#include "gazebo/gui/building/BuildingEditorEvents.hh"
 #include "gazebo/gui/building/BuildingMaker.hh"
 #include "gazebo/gui/building/BuildingModelManip.hh"
 
@@ -28,6 +29,11 @@ using namespace gui;
 BuildingModelManip::BuildingModelManip()
 {
   this->parent = NULL;
+  this->level = 0;
+
+  this->connections.push_back(
+  gui::editor::Events::ConnectChangeBuildingLevel(
+    boost::bind(&BuildingModelManip::OnChangeLevel, this, _1)));
 }
 
 /////////////////////////////////////////////////
@@ -58,6 +64,18 @@ std::string BuildingModelManip::GetName() const
 rendering::VisualPtr BuildingModelManip::GetVisual() const
 {
   return this->visual;
+}
+
+/////////////////////////////////////////////////
+common::Color BuildingModelManip::GetColor() const
+{
+  return this->color;
+}
+
+/////////////////////////////////////////////////
+std::string BuildingModelManip::GetTexture() const
+{
+  return this->texture;
 }
 
 /////////////////////////////////////////////////
@@ -263,6 +281,30 @@ void BuildingModelManip::OnRotationChanged(double _roll, double _pitch,
 }
 
 /////////////////////////////////////////////////
+void BuildingModelManip::OnLevelChanged(int _level)
+{
+  this->SetLevel(_level);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::OnColorChanged(QColor _color)
+{
+  this->SetColor(_color);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::OnTextureChanged(QString _texture)
+{
+  this->SetTexture(_texture);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::OnTransparencyChanged(float _transparency)
+{
+  this->SetTransparency(_transparency);
+}
+
+/////////////////////////////////////////////////
 void BuildingModelManip::OnDeleted()
 {
   this->maker->RemovePart(this->name);
@@ -313,4 +355,55 @@ void BuildingModelManip::SetSize(double _width, double _depth, double _height)
       - math::Vector3(dScale.x/2.0, dScale.y/2.0, dScale.z/2.0);
 
   this->visual->SetPosition(newPos);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::SetColor(QColor _color)
+{
+  common::Color newColor(_color.red(), _color.green(), _color.blue());
+  this->color = newColor;
+  this->visual->GetParent()->SetAmbient(this->color);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::SetTexture(QString _texture)
+{
+  // TODO For now setting existing material scripts.
+  // Add support for custom textures.
+  this->texture = "Gazebo/Grey";
+  if (_texture == ":wood.jpg")
+    this->texture = "Gazebo/Wood";
+  else if (_texture == ":tiles.jpg")
+    this->texture = "Gazebo/CeilingTiled";
+
+  this->visual->GetParent()->SetMaterial(this->texture);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::SetTransparency(float _transparency)
+{
+  this->visual->GetParent()->SetTransparency(_transparency);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::SetLevel(const int _level)
+{
+  this->level = _level;
+}
+
+/////////////////////////////////////////////////
+int BuildingModelManip::GetLevel() const
+{
+  return this->level;
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::OnChangeLevel(int _level)
+{
+  if (this->level > _level)
+    this->SetTransparency(1.0);
+  else if (this->level < _level)
+    this->SetTransparency(0.0);
+  else
+    this->SetTransparency(0.4);
 }