+++ gazebo/gui/building/EditorView.cc
@@ -253,50 +253,44 @@ void EditorView::mousePressEvent(QMouseEvent *_event)
 /////////////////////////////////////////////////
 void EditorView::mouseReleaseEvent(QMouseEvent *_event)
 {
-  switch (this->drawMode)
+  if (this->drawMode == WALL)
   {
-    case NONE:
-      break;
-    case WALL:
-      this->DrawWall(_event->pos());
-      break;
-    case WINDOW:
-      if (this->drawInProgress)
+    this->DrawWall(_event->pos());
+  }
+  else if (this->drawMode != NONE)
+  {
+    if (this->drawInProgress)
+    {
+      if (this->drawMode == WINDOW)
       {
         this->windowList.push_back(dynamic_cast<WindowItem *>(
             this->currentMouseItem));
-        this->drawMode = NONE;
-        this->drawInProgress = false;
-        gui::editor::Events::createBuildingEditorItem(std::string());
       }
-      break;
-    case DOOR:
-      if (this->drawInProgress)
+      else if (this->drawMode == DOOR)
       {
         this->doorList.push_back(dynamic_cast<DoorItem *>(
             this->currentMouseItem));
-        this->drawMode = NONE;
-        this->drawInProgress = false;
-        gui::editor::Events::createBuildingEditorItem(std::string());
       }
-      break;
-    case STAIRS:
-      if (this->drawInProgress)
+      else if (this->drawMode == STAIRS)
       {
-        this->stairsList.push_back(dynamic_cast<StairsItem *>(
-            this->currentMouseItem));
+        StairsItem *stairsItem = dynamic_cast<StairsItem *>(
+            this->currentMouseItem);
+        stairsItem->Set3dTexture(QString(""));
+        stairsItem->Set3dColor(Qt::white);
+        this->stairsList.push_back(stairsItem);
         if ((this->currentLevel) < static_cast<int>(floorList.size()))
         {
-          EditorItem *item = dynamic_cast<EditorItem *>(this->currentMouseItem);
-          this->buildingMaker->AttachManip(this->itemToVisualMap[item],
+          this->buildingMaker->AttachManip(this->itemToVisualMap[stairsItem],
               this->itemToVisualMap[floorList[this->currentLevel]]);
         }
-        this->drawMode = NONE;
-        this->drawInProgress = false;
-        gui::editor::Events::createBuildingEditorItem(std::string());
       }
-    default:
-      break;
+      dynamic_cast<EditorItem *>(this->currentMouseItem)->
+          SetHighlighted(false);
+
+      this->drawMode = NONE;
+      this->drawInProgress = false;
+      gui::editor::Events::createBuildingEditorItem(std::string());
+    }
   }
 
   if (!this->drawInProgress)
@@ -327,7 +321,7 @@ void EditorView::mouseMoveEvent(QMouseEvent *_event)
         QPointF pf;
         pf = p2;
 
-        if (QApplication::keyboardModifiers() & Qt::ShiftModifier)
+        if (!(QApplication::keyboardModifiers() & Qt::ShiftModifier))
         {
           double distanceToClose = 30;
 
@@ -363,20 +357,27 @@ void EditorView::mouseMoveEvent(QMouseEvent *_event)
             }
           }
 
-          // Snap to 15 degrees increments
           if (!this->snapToGrabber)
           {
+            // Snap to angular increments
             QLineF newLine(p1, p2);
             double angle = GZ_DTOR(QLineF(p1, p2).angle());
-            double range = GZ_DTOR(15);
-            int increment = angle / range;
+            double range = GZ_DTOR(SegmentItem::SnapAngle);
+            int angleIncrement = angle / range;
 
-            if ((angle - range*increment) > range*0.5)
-              increment++;
-            angle = -range*increment;
+            if ((angle - range*angleIncrement) > range*0.5)
+              angleIncrement++;
+            angle = -range*angleIncrement;
 
-            pf.setX(p1.x() + qCos(angle)*newLine.length());
-            pf.setY(p1.y() + qSin(angle)*newLine.length());
+            // Snap to length increments
+            double newLength = newLine.length();
+            double lengthIncrement = SegmentItem::SnapLength /
+                wallSegmentItem->GetScale();
+            newLength  = round(newLength/lengthIncrement)*lengthIncrement-
+                wallSegmentItem->GetThickness();
+
+            pf.setX(p1.x() + qCos(angle)*newLength);
+            pf.setY(p1.y() + qSin(angle)*newLength);
           }
         }
         wallSegmentItem->SetEndPoint(pf);
@@ -429,6 +430,8 @@ void EditorView::mouseMoveEvent(QMouseEvent *_event)
         if (distance > 30 || t > 1.0 || t < 0.0)
         {
           editorItem->setParentItem(NULL);
+          editorItem->SetPositionOnWall(0);
+          editorItem->SetAngleOnWall(0);
           this->buildingMaker->DetachManip(this->itemToVisualMap[editorItem],
                 this->itemToVisualMap[wallSegmentItem]);
           editorItem->SetRotation(editorItem->GetRotation()
@@ -566,6 +569,8 @@ void EditorView::DeleteItem(EditorItem *_item)
   if (!_item)
     return;
 
+  this->buildingMaker->DetachAllChildren(this->itemToVisualMap[_item]);
+
   if (_item->GetType() == "WallSegment")
   {
     WallSegmentItem *wallSegmentItem = dynamic_cast<WallSegmentItem *>(_item);
@@ -618,6 +623,43 @@ void EditorView::DrawWall(const QPoint &_pos)
 
     wallSegmentItem = new WallSegmentItem(pointStart, pointEnd,
         this->levelDefaultHeight);
+
+    if (!(QApplication::keyboardModifiers() & Qt::ShiftModifier))
+    {
+      double distanceToClose = 30;
+
+      // Snap to other walls' points
+      QList<QGraphicsItem *> itemsList = this->scene()->items(QRectF(
+          QPointF(pointStart.x() - distanceToClose/2,
+                  pointStart.y() - distanceToClose/2),
+          QPointF(pointStart.x() + distanceToClose/2,
+                  pointStart.y() + distanceToClose/2)));
+      for (QList<QGraphicsItem *>::iterator it = itemsList.begin();
+          it  != itemsList.end(); ++it)
+      {
+        WallSegmentItem *anotherWall = dynamic_cast<WallSegmentItem *>(*it);
+        if (anotherWall)
+        {
+          if (QVector2D(pointStart - anotherWall->GetStartPoint()).length()
+              <= distanceToClose)
+          {
+            wallSegmentItem->SetStartPoint(anotherWall->GetStartPoint());
+            this->LinkGrabbers(anotherWall->grabbers[0],
+                wallSegmentItem->grabbers[0]);
+            break;
+          }
+          else if (QVector2D(pointStart - anotherWall->GetEndPoint()).length()
+              <= distanceToClose)
+          {
+            wallSegmentItem->SetStartPoint(anotherWall->GetEndPoint());
+            this->LinkGrabbers(anotherWall->grabbers[1],
+                wallSegmentItem->grabbers[0]);
+            break;
+          }
+        }
+      }
+    }
+
     wallSegmentItem->SetLevel(this->currentLevel);
     wallSegmentItem->SetLevelBaseHeight(this->levels[this->currentLevel]->
         baseHeight);
@@ -641,9 +683,9 @@ void EditorView::DrawWall(const QPoint &_pos)
     this->snapGrabberCurrent = NULL;
 
     wallSegmentItem = dynamic_cast<WallSegmentItem*>(this->currentMouseItem);
-    // Select -> deselect to trigger change
-    wallSegmentItem->setSelected(true);
-    wallSegmentItem->setSelected(false);
+    wallSegmentItem->Set3dTexture(QString(""));
+    wallSegmentItem->Set3dColor(Qt::white);
+    wallSegmentItem->SetHighlighted(false);
     wallSegmentList.push_back(wallSegmentItem);
     if (wallSegmentItem->GetLevel() > 0)
     {
@@ -953,6 +995,7 @@ void EditorView::OnAddLevel()
   newLevel->baseHeight = maxHeight;
 
   FloorItem *floorItem = new FloorItem();
+  this->levels[this->currentLevel]->floorItem = floorItem;
   std::vector<WallSegmentItem *> newWalls;
   std::map<WallSegmentItem *, WallSegmentItem *> clonedWallMap;
   for (std::vector<WallSegmentItem *>::iterator it = wallSegmentList.begin();
@@ -980,10 +1023,9 @@ void EditorView::OnAddLevel()
     this->itemToVisualMap[wallSegmentItem] = wallSegmentName;
 
     floorItem->AttachWallSegment(wallSegmentItem);
-
-    // Select -> deselect to trigger change
-    wallSegmentItem->setSelected(true);
-    wallSegmentItem->setSelected(false);
+    wallSegmentItem->Set3dTexture(QString(""));
+    wallSegmentItem->Set3dColor(Qt::white);
+    wallSegmentItem->SetHighlighted(false);
   }
 
   // Clone linked grabber relations
@@ -1027,6 +1069,9 @@ void EditorView::OnAddLevel()
   this->itemToVisualMap[floorItem] = floorName;
   this->scene()->addItem(floorItem);
   this->floorList.push_back(floorItem);
+  floorItem->Set3dTexture(QString(""));
+  floorItem->Set3dColor(Qt::white);
+  floorItem->SetHighlighted(false);
 }
 
 /////////////////////////////////////////////////
@@ -1042,6 +1087,7 @@ void EditorView::DeleteLevel(int _level)
       || _level >= static_cast<int>(this->levels.size()))
     return;
 
+  // Delete current level and move to level below or above
   int newLevelIndex = _level - 1;
   if (newLevelIndex < 0)
     newLevelIndex = _level + 1;
@@ -1168,6 +1214,18 @@ void EditorView::OnChangeLevel(int _level)
 void EditorView::OnOpenLevelInspector()
 {
   this->levelInspector->SetLevelName(this->levels[this->currentLevel]->name);
+  FloorItem *floorItem = this->levels[this->currentLevel]->floorItem;
+  if (floorItem)
+  {
+    this->levelInspector->floorWidget->show();
+    this->levelInspector->SetFloorColor(floorItem->Get3dColor());
+    this->levelInspector->SetFloorTexture(floorItem->Get3dTexture());
+  }
+  else
+  {
+    this->levelInspector->floorWidget->hide();
+  }
+  this->levelInspector->move(QCursor::pos());
   this->levelInspector->show();
 }
 
@@ -1179,6 +1237,14 @@ void EditorView::OnLevelApply()
 
   std::string newLevelName = dialog->GetLevelName();
   this->levels[this->currentLevel]->name = newLevelName;
+  FloorItem *floorItem = this->levels[this->currentLevel]->floorItem;
+  if (floorItem)
+  {
+    floorItem->Set3dTexture(dialog->GetFloorTexture());
+    floorItem->Set3dColor(dialog->GetFloorColor());
+    floorItem->Set3dTransparency(0.4);
+    floorItem->FloorChanged();
+  }
   gui::editor::Events::updateLevelWidget(this->currentLevel, newLevelName);
 }