+++ gazebo/gui/building/WallSegmentItem.hh
@@ -19,6 +19,7 @@
 #define _WALL_SEGMENT_ITEM_HH_
 
 #include "gazebo/gui/qt.h"
+#include "gazebo/gui/building/MeasureItem.hh"
 #include "gazebo/gui/building/SegmentItem.hh"
 #include "gazebo/gui/building/BuildingItem.hh"
 
@@ -72,6 +73,9 @@ namespace gazebo
       /// \brief Update inspector with current values.
       public: void UpdateInspector();
 
+      // Documentation inherited
+      public: void SetHighlighted(bool _highlighted);
+
       /// \brief Update wall segment when segment updated.
       protected: void SegmentUpdated();
 
@@ -105,8 +109,8 @@ namespace gazebo
       /// \brief Height of the wall segment in meters.
       private: double wallHeight;
 
-      /// \brief Scale for converting pixels to metric units.
-      private: double scale;
+      /// \brief This wall segment's measure item.
+      private: MeasureItem *measure;
 
       /// \brief Qt action for opening the inspector.
       private: QAction *openInspectorAct;