+++ gazebo/gui/model/ModelCreator.hh
@@ -161,6 +161,17 @@ namespace gazebo
       /// \param[in] _checked True if the menu item is checked
       private slots: void OnEdit(bool _checked);
 
+      /// \brief Qt callback when the copy action is triggered.
+      private slots: void OnCopy();
+
+      /// \brief Qt callback when the paste action is triggered.
+      private slots: void OnPaste();
+
+      /// \brief Mouse event filter callback when mouse is pressed.
+      /// \param[in] _event The mouse event.
+      /// \return True if the event was handled
+      private: bool OnMousePress(const common::MouseEvent &_event);
+
       /// \brief Mouse event filter callback when mouse is released.
       /// \param[in] _event The mouse event.
       /// \return True if the event was handled
@@ -198,6 +209,19 @@ namespace gazebo
       /// \brief Get a template SDF string of a simple model.
       /// \return Template SDF string of a simple model.
       private: std::string GetTemplateSDFString();
+
+      /// \brief Callback when a specific alignment configuration is set.
+      /// \param[in] _axis Axis of alignment: x, y, or z.
+      /// \param[in] _config Configuration: min, center, or max.
+      /// \param[in] _target Target of alignment: first or last.
+      /// \param[in] _bool True to preview alignment without publishing
+      /// to server.
+      private: void OnAlignMode(const std::string &_axis,
+          const std::string &_config, const std::string &_target,
+          bool _preview);
+
+      /// \brief Deselect all currently selected visuals.
+      private: void DeselectAll();
 
       /// \brief Qt callback when a delete signal has been emitted.
       /// \param[in] _name Name of the part or model to delete.
@@ -248,6 +272,9 @@ namespace gazebo
       /// \brief Counter for generating a unique model name.
       private: int modelCounter;
 
+      /// \brief Transparency value for model being edited.
+      private: double editTransparency;
+
       /// \brief Type of part being added.
       private: PartType addPartType;
 
@@ -271,8 +298,14 @@ namespace gazebo
       /// \brief origin of the model.
       private: math::Pose origin;
 
-      /// \brief Selected partv visual;
-      private: rendering::VisualPtr selectedVis;
+      /// \brief A list of selected visuals.
+      private: std::vector<rendering::VisualPtr> selectedVisuals;
+
+      /// \brief Names of parts copied through g_copyAct
+      private: std::vector<std::string> copiedPartNames;
+
+      /// \brief The last mouse event
+      private: common::MouseEvent lastMouseEvent;
     };
     /// \}
 
@@ -323,10 +356,10 @@ namespace gazebo
       /// \brief Pose of part.
       public: math::Pose pose;
 
-      /// \brief Name of part.
+      /// \brief Inertial properties of part.
       public: physics::Inertial *inertial;
 
-      /// \brief Name of part.
+      /// \brief Collision properties of part.
       public: std::vector<physics::CollisionPtr> collisions;
 
       /// \brief Sensor data