+++ gazebo/gui/model/ModelEditor.cc
@@ -37,7 +37,7 @@ ModelEditor::ModelEditor(MainWindow *_mainWindow)
   this->modelPalette = new ModelEditorPalette(_mainWindow);
   this->Init("modelEditorTab", "Model Editor", this->modelPalette);
 
-  connect(g_editModelAct, SIGNAL(triggered(bool)), this, SLOT(OnEdit(bool)));
+  connect(g_editModelAct, SIGNAL(toggled(bool)), this, SLOT(OnEdit(bool)));
 
   this->connections.push_back(
       gui::model::Events::ConnectFinishModel(
@@ -62,16 +62,16 @@ void ModelEditor::OnEdit(bool /*_checked*/)
     this->mainWindow->ShowLeftColumnWidget();
     this->mainWindow->Play();
   }
-  event::Events::setSelectedEntity("", "normal");
   this->active = !this->active;
   this->ToggleToolbar();
-  g_editModelAct->setChecked(this->active);
+//  g_editModelAct->setChecked(this->active);
 }
 
 /////////////////////////////////////////////////
 void ModelEditor::OnFinish()
 {
-  this->OnEdit(g_editModelAct->isChecked());
+//  this->OnEdit(g_editModelAct->isChecked());
+  g_editModelAct->trigger();
 }
 
 /////////////////////////////////////////////////
@@ -86,10 +86,10 @@ void ModelEditor::ToggleToolbar()
         actions[i] == g_rotateAct ||
         actions[i] == g_translateAct ||
         actions[i] == g_scaleAct ||
-        actions[i] == g_screenshotAct)
-//        actions[i] == g_copyAct -- issue #1314
-//        actions[i] == g_pasteAct
-//        align tool              -- issue #1323
+        actions[i] == g_screenshotAct ||
+        actions[i] == g_copyAct ||
+        actions[i] == g_pasteAct ||
+        actions[i] == g_alignButtonAct)
 //        actions[i] == g_snapAct -- issue #1318
     {
       actions[i]->setVisible(true);